Figure editor is responsive. Began work on reading/writing figure editor data to forms.
Initial approach is to encode editor data as a flat string,
e.g. "neighbor balance+allemande_right 540 degrees for 8", and decode
it using (supported?) regexp facility in Javascript. However, a better
idea now occurs: write in JSON. So I'm saving this mostly-working
string reader/writer to git, and tearing it down, and making a JSON
version.